Version 1.34n
      + Non-rectangular selections can now be dragged paste the edges of   
        the image.                                                         
      + Added the getDateAndTime() macro function.                         
      + The command recorder correctly records threshold levels of signed  
        16-bit images.                                                     
      + Fixed a bug that prevented filtering of images using the "3-3-2    
        RGB" LUT.                                                          
      + Fixed a bug that caused "Standard Deviation" projections using     
        Image/Stacks/Z Project to fail with some 16-bit stacks.            
      + Fixed a bug with Process/Binary/Threshold that caused all slice    
        labels of the output stack to be set to the first slice label of   
        the input stack.                                                   
      + Fixed a bug that caused drawn lines to be the same width when the  
        line width was set to both w and w+1.                              
      + Fixed a bug that caused drawn lines to be positioned incorrectly   
        when the starting coordinates were negative.                       
      + Fixed a bug that caused batch mode macros doing 90 degree rotations
        to fail.                                                           
      + Fixed a bug that caused the lengthOf() macro function to fail when 
        called with a string array element.                                
      + Fixed a bug that caused the exit() macro function to fail when     
        called with a variable as an argument.                             

17 May 2005

+++ ij/plugin/ZProjector.java
@@ -553,7 +553,7 @@ public class ZProjector implements PlugIn {
 		}
 	
 		public void projectSlice(short[] pixels) {
-			int v;
+			float v;
 		    for(int i=0; i<len; i++) {
 		    	v = pixels[i]&0xffff;
 				sum[i] += v;
